% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cov_random_slope}
\alias{cov_random_slope}
\title{Within-subject covariance matrix for the random-intercept,
random-slope structure}
\usage{
cov_random_slope(times, tau_0, tau_1, tau_01, sigma)
}
\arguments{
\item{times}{Numeric vector of visit times.}

\item{tau_0}{Numeric, SD of the random intercept.}

\item{tau_1}{Numeric, SD of the random slope.}

\item{tau_01}{Numeric, covariance between random intercept and slope.}

\item{sigma}{Numeric, residual SD.}
}
\value{
A k x k covariance matrix.
}
\description{
Within-subject covariance matrix for the random-intercept,
random-slope structure
}
